some styles are broken?
