--- docs/basic-demo.mdx
@@ -1,38 +1,0 @@
----
-name: Basic
-menu: Demo
----
-
-import {Playground} from 'docz';
-import Collapse from '../src/Collapse';
-import {DemoContainer, CollapseDemo} from './components/ui-components';
-
-# Basic Demo
-
-<Playground>
-  {/* Fixed height to minimize the code snippet jumping around :) */}
-  <DemoContainer>
-    <Collapse defaultOpen>
-      {({getCollapseProps, getTogglerProps, isOpen}) => (
-        <React.Fragment>
-          <button {...getTogglerProps({style: {marginBottom: '1em'}})}>
-            {isOpen ? 'Close' : 'Open'}
-          </button>
-          <div {...getCollapseProps()}>
-            <CollapseDemo>
-              In the morning I walked down the Boulevard to the rue Soufflot for
-              coffee and brioche. It was a fine morning. The horse-chestnut
-              trees in the Luxembourg gardens were in bloom. There was the
-              pleasant early-morning feeling of a hot day. I read the papers
-              with the coffee and then smoked a cigarette. The flower-women were
-              coming up from the market and arranging their daily stock.
-              Students went by going up to the law school, or down to the
-              Sorbonne. The Boulevard was busy with trams and people going to
-              work.
-            </CollapseDemo>
-          </div>
-        </React.Fragment>
-      )}
-    </Collapse>
-  </DemoContainer>
-</Playground>
