+++ docs/basic/basic-demo.mdx
@@ -1,0 +1,11 @@
+---
+name: Basic
+menu: Demo
+---
+
+import {Playground} from 'docz';
+import Demo from './basic-demo';
+
+# Basic Demo
+
+<Demo />
