--- docs/controlled-demo.mdx
@@ -1,58 +1,0 @@
----
-title: Controlled
-route: /controlled-demo
-menu: Demo
----
-
-import ControlledDemo from './controlled-demo';
-import {DemoContainer, CollapseDemo} from './components/ui-components';
-
-# Controlled
-
-While ReactCollapsed can control the expanded behavior for you, you're welcome to handle it yourself via the `isOpen` prop.
-
-<ControlledDemo />
-
-```jsx
-class ControlledDemo extends React.Component {
-  state = {
-    isOpen: true
-  };
-
-  handleClick = () => this.setState(({isOpen}) => ({isOpen: !isOpen}));
-
-  render() {
-    const {isOpen} = this.state;
-    return (
-      <DemoContainer>
-        <Collapse defaultOpen isOpen={isOpen}>
-          {({getCollapseProps, getTogglerProps}) => (
-            <React.Fragment>
-              <button
-                {...getTogglerProps({
-                  onClick: this.handleClick
-                })}
-              >
-                {isOpen ? 'Close' : 'Open'}
-              </button>
-              <div {...getCollapseProps()}>
-                <p>
-                  In the morning I walked down the Boulevard to the rue Soufflot
-                  for coffee and brioche. It was a fine morning. The
-                  horse-chestnut trees in the Luxembourg gardens were in bloom.
-                  There was the pleasant early-morning feeling of a hot day. I
-                  read the papers with the coffee and then smoked a cigarette.
-                  The flower-women were coming up from the market and arranging
-                  their daily stock. Students went by going up to the law
-                  school, or down to the Sorbonne. The Boulevard was busy with
-                  trams and people going to work.
-                </p>
-              </div>
-            </React.Fragment>
-          )}
-        </Collapse>
-      </DemoContainer>
-    );
-  }
-}
-```
