--- docs/hooks-demo.mdx
@@ -1,13 +1,0 @@
----
-name: Hooks
-menu: Demo
----
-
-import {Playground} from 'docz';
-import Demo from './hooks-demo';
-import NestedDemo from './nested-hooks-demo';
-
-# Hooks Demo
-
-<Demo />
-<NestedDemo />
