+++ docs/nested/nested-demo.mdx
@@ -1,0 +1,11 @@
+---
+name: Nested
+route: /nested-demo
+menu: Demo
+---
+
+import Demo from './nested-demo';
+
+# Nested Demo
+
+<Demo />
