+++ docs/nested/nested-hooks-demo.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import {useCollapse} from '../src/collapsed';
-import {CollapseDemo} from './components/ui-components';
+import {useCollapse} from '../../src/collapsed';
+import {CollapseDemo} from '../_ui-components/ui-components';
 
 export default function NestedDemo() {
   const {
