--- docs/transition-styles-demo.mdx
@@ -1,46 +1,0 @@
----
-name: Custom In-And-Out
-menu: Demo
----
-
-import {Playground} from 'docz';
-import Collapse from '../src/Collapse';
-import {DemoContainer, CollapseDemo} from './components/ui-components';
-
-# Basic Demo
-
-<Playground>
-  {/* Fixed height to minimize the code snippet jumping around :) */}
-  <DemoContainer>
-    <Collapse
-      defaultOpen
-      duration={{in: 900, out: 500}}
-      easing={{
-        in: 'cubic-bezier(0.470, 0.000, 0.745, 0.715)',
-        out: 'cubic-bezier(0.230, 1.000, 0.320, 1.000)'
-      }}
-      delay={{out: 100}}
-    >
-      {({getCollapseProps, getTogglerProps, isOpen}) => (
-        <React.Fragment>
-          <button {...getTogglerProps({style: {marginBottom: '1em'}})}>
-            {isOpen ? 'Close' : 'Open'}
-          </button>
-          <div {...getCollapseProps()}>
-            <CollapseDemo>
-              In the morning I walked down the Boulevard to the rue Soufflot for
-              coffee and brioche. It was a fine morning. The horse-chestnut
-              trees in the Luxembourg gardens were in bloom. There was the
-              pleasant early-morning feeling of a hot day. I read the papers
-              with the coffee and then smoked a cigarette. The flower-women were
-              coming up from the market and arranging their daily stock.
-              Students went by going up to the law school, or down to the
-              Sorbonne. The Boulevard was busy with trams and people going to
-              work.
-            </CollapseDemo>
-          </div>
-        </React.Fragment>
-      )}
-    </Collapse>
-  </DemoContainer>
-</Playground>
