+++ docs/transition-styles/transition-styles-demo.mdx
@@ -1,0 +1,11 @@
+---
+name: Custom In-And-Out
+route: /transition-styles
+menu: Demo
+---
+
+import Demo from './transition-styles-demo';
+
+# Transition Styles Demo
+
+<Demo />
